# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "AceScape API",
      "version": "1.0.0",
      "description": "Documentation de l'API pour AceScape"
    },
    "paths": {
      "/api/login": {
        "post": {
          "summary": "Authentifie un utilisateur",
          "tags": ["User"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": ["username", "password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Authentification réussie",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "401": {
              "description": "Nom d'utilisateur ou mot de passe incorrect"
            },
            "500": {
              "description": "Erreur interne du serveur"
            }
          }
        }
      },
      "/api/users": {
        "post": {
          "summary": "Ajoute un utilisateur",
          "tags": ["User"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewUser"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Utilisateur créé",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "Ressource non trouvée"
            },
            "405": {
              "description": "Erreur de validation ou duplication"
            }
          }
        },
        "get": {
          "summary": "Recherche plusieurs utilisateurs avec pagination",
          "tags": ["User"],
          "parameters": [
            {
              "in": "query",
              "name": "page",
              "schema": {
                "type": "integer"
              },
              "description": "Numéro de page"
            },
            {
              "in": "query",
              "name": "pageSize",
              "schema": {
                "type": "integer"
              },
              "description": "Taille de la page"
            },
            {
              "in": "query",
              "name": "q",
              "schema": {
                "type": "string"
              },
              "description": "Valeur de recherche"
            }
          ],
          "responses": {
            "200": {
              "description": "Utilisateurs trouvés",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Erreur de validation"
            },
            "500": {
              "description": "Erreur serveur"
            }
          }
        }
      },
      "/api/users/multiple": {
        "post": {
          "summary": "Ajoute plusieurs utilisateurs",
          "tags": ["User"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewUser"
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Utilisateurs créés",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Erreur de validation"
            }
          }
        },
        "get": {
          "summary": "Recherche plusieurs utilisateurs par ID",
          "tags": ["User"],
          "parameters": [
            {
              "in": "query",
              "name": "id",
              "required": true,
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "description": "Liste des IDs des utilisateurs"
            }
          ],
          "responses": {
            "200": {
              "description": "Utilisateurs trouvés",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Utilisateur(s) non trouvé(s)"
            },
            "405": {
              "description": "Erreur de validation"
            },
            "500": {
              "description": "Erreur serveur"
            }
          }
        },
        "put": {
          "summary": "Modifie plusieurs utilisateurs",
          "tags": ["User"],
          "parameters": [
            {
              "in": "query",
              "name": "id",
              "required": true,
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "description": "Liste des IDs des utilisateurs"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Utilisateurs modifiés",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Utilisateur(s) non trouvé(s)"
            },
            "405": {
              "description": "Erreur de validation ou duplication"
            },
            "500": {
              "description": "Erreur serveur"
            }
          }
        },
        "delete": {
          "summary": "Supprime plusieurs utilisateurs",
          "tags": ["User"],
          "parameters": [
            {
              "in": "query",
              "name": "id",
              "required": true,
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "description": "Liste des IDs des utilisateurs"
            }
          ],
          "responses": {
            "200": {
              "description": "Utilisateurs supprimés"
            },
            "404": {
              "description": "Utilisateur(s) non trouvé(s)"
            },
            "405": {
              "description": "Erreur de validation"
            },
            "500": {
              "description": "Erreur serveur"
            }
          }
        }
      },
      "/api/users/{id}": {
        "get": {
          "summary": "Recherche un utilisateur par ID",
          "tags": ["User"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "ID de l'utilisateur"
            }
          ],
          "responses": {
            "200": {
              "description": "Utilisateur trouvé",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "Utilisateur non trouvé"
            },
            "405": {
              "description": "Erreur de validation"
            },
            "500": {
              "description": "Erreur serveur"
            }
          }
        },
        "put": {
          "summary": "Modifie un utilisateur",
          "tags": ["User"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "ID de l'utilisateur"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Utilisateur modifié",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "Utilisateur non trouvé"
            },
            "405": {
              "description": "Erreur de validation ou duplication"
            },
            "500": {
              "description": "Erreur serveur"
            }
          }
        },
        "delete": {
          "summary": "Supprime un utilisateur",
          "tags": ["User"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "ID de l'utilisateur"
            }
          ],
          "responses": {
            "200": {
              "description": "Utilisateur supprimé"
            },
            "404": {
              "description": "Utilisateur non trouvé"
            },
            "405": {
              "description": "Erreur de validation"
            },
            "500": {
              "description": "Erreur serveur"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "required": ["id", "name", "email"]
        },
        "NewUser": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          },
          "required": ["name", "email", "password"]
        }
      }
    }
  }
  